master - Method References Between Components
This chapter explained how event callbacks can be provided to child
components. This is important as it gives stateless functional components the ability
to execute callbacks that update state in parents class based
components. There are two methods of passing callback paramters to child
components. The first is through the use of bind(..) and the second is
using by passing a lambda expression that is executed when an event
occurs. The bind(..) method is preferred as it has less overhead.