master - Working with Inline Styles
This chapter explained how to apply component styles using inline styles
defined as JavaScript objects. Inline styles are scopped to the
component they're applied to but have some limitations. There is a third
way of defining styles that is a mix between inline and stylesheet
styles that is the preferred method of providing styles to components.