section-4 - Lists & Keys
This chapter stressed the importance of providing keys to JSX components
in lists. It's important to assign keys to list components due to the
way ReactJS analyzes the DOM when determining what components to update.
By providing a key, it allows ReactJS to render only whats changed and
not the entire list.