section-7 - Using setState Correctly
This lecture explained how setState is actually an asynchronous
function. Because setState is an asynchronous function, referencing old state
will cause race conditions. The proper way to invoke setState is by
passing it a function that accepts the old state and current properties
as arguments. This allows the call to setState to reference previous
state.