section-7 - Using References ("ref")
This lecture explained how to reference elements/components within a
stateful component by using the, "ref" property. The ref property makes
it possible to add a reference to an element on the parent component.
Once added, elements can be referenced one render() is invoked. A good
place to do this is within componentDidMount().