section-7 - More on the React ref API (16.3+)
This lecture explained how to use React.createRef() to create
references. This method creates a reference that can then be provided to
the ref attribute on a component. This is different from how it was done
in the past by the fact that lambda expressions don't have to be created
to explicitly set the reference.