section-7 - The Context API (React 16.3+)
This lecture introduced the Context API. The context API is a new API
introduced to allow passing global state around an application without
having to supply it through properties. This is done by using
React.createContext(..) and using the created subcomponents from the
generated component called Provider and Consumer. Using props is the
prefered method of passing around most state, but this works well for
global state.